+++ components/outfits/deleteButton.tsx
@@ -10,10 +10,13 @@ type Props = {
 
 export default function DeleteOutfitButton(props: Props) {
   const router = useRouter();
+
   async function handleDelete() {
+    if (!confirm("Are you sure you want to delete this outfit?")) return;
     await deleteOutfit(props.id);
     router.push("/outfits");
   }
+
   return (
     <Button variant="destructive" onClick={handleDelete}>
       Delete item
